成本导出月份格式化

+++ erp-model/erp-model-dmp/src/main/java/com/erp/model/dmp/dto/DmpOrderInfoExcelDTO.java
@@ -175,6 +175,6 @@ public class DmpOrderInfoExcelDTO implements Serializable {
      * 销售额[原币别]
      */
     @ExcelProperty(value = "销售额[原币别]", index = 22)
-    private Integer sellAmount;
+    private BigDecimal sellAmount;
 
 }
